1. implemented fragment views 2. fixed signin page activity finish to properly finish activity before starting another activity. 3. Added menu bar with signout option.
