+++ .idea/modules.xml
@@ -3,7 +3,7 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
-      <module fileurl="file://$PROJECT_DIR$/cremeBrulee.iml" filepath="$PROJECT_DIR$/cremeBrulee.iml" />
+      <module fileurl="file://$PROJECT_DIR$/cremebrulee.iml" filepath="$PROJECT_DIR$/cremebrulee.iml" />
     </modules>
   </component>
 </project>
