+++ app/src/main/res/layout/activity_profile.xml
@@ -8,7 +8,7 @@
     android:background="@color/light_blue"
 
 
-    tools:context="plscuddleme.yuhanlee.cremebrulee.MainActivity">
+    tools:context="plscuddleme.yuhanlee.cremebrulee.Tabs">
 
 
     <RelativeLayout
@@ -35,13 +35,6 @@
                     android:cropToPadding="false"
                     android:scaleType="centerCrop"
                     android:src="@drawable/dog3" />
-
-                <Button
-                    android:id="@+id/log_out"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="logout tmp button" />
-
 
 
                 <ImageView
