+++ app/src/main/res/layout/activity_sign_in.xml
@@ -67,13 +67,6 @@
             android:gravity="center"
             android:text="@string/join_now" />
 
-        <Button
-            android:id="@+id/button1"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="40px"
-            android:text="Go to tabs" />
-
 
     </LinearLayout>
 </RelativeLayout>
