+++ app/src/main/res/layout/activity_tabs.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="plscuddleme.yuhanlee.cremebrulee.tabs">
+    tools:context="plscuddleme.yuhanlee.cremebrulee.Tabs">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/appbar"
