+++ app/src/main/res/layout/fragment_tabs.xml
@@ -4,7 +4,7 @@
     android:id="@+id/constraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="plscuddleme.yuhanlee.cremebrulee.tabs$PlaceholderFragment">
+    tools:context="plscuddleme.yuhanlee.cremebrulee.Tabs$PlaceholderFragment">
 
     <TextView
         android:id="@+id/section_label"
