+++ app/src/main/res/values/strings.xml
@@ -13,7 +13,9 @@
     <string name="signIn_here">Have an account? Sign in Here</string>
     <string name="register">Register</string>
     <string name="title_activity_tabs">tabs</string>
-    <string name="settings">Settings</string>
+    <string name="logout">logout</string>
+    <string name="settings">settings</string>
+
     <string name="section_format">Hello World from section: </string>
